+++ src/main/java/com/back/domain/chat/chat/entity/ChatMember.java
@@ -1,6 +1,6 @@
 package com.back.domain.chat.chat.entity;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
