+++ src/main/java/com/back/domain/chat/chat/entity/ChatRoom.java
@@ -1,6 +1,6 @@
 package com.back.domain.chat.chat.entity;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 import com.back.domain.post.post.entity.Post;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.*;
