+++ src/main/java/com/back/domain/chat/chat/repository/ChatRoomRepositoryImpl.java
@@ -17,7 +17,7 @@ import java.util.Optional;
 
 import static com.back.domain.chat.chat.entity.QChatMember.chatMember;
 import static com.back.domain.chat.chat.entity.QChatRoom.chatRoom;
-import static com.back.domain.member.member.entity.QMember.member;
+import static com.back.domain.member.entity.QMember.member;
 import static com.back.domain.post.post.entity.QPost.post;
 
 @RequiredArgsConstructor
