+++ src/main/java/com/back/domain/chat/chat/service/ChatService.java
@@ -4,8 +4,8 @@ import com.back.domain.chat.chat.dto.ChatRoomDto;
 import com.back.domain.chat.chat.dto.CreateChatRoomResBody;
 import com.back.domain.chat.chat.entity.ChatRoom;
 import com.back.domain.chat.chat.repository.ChatRoomRepository;
-import com.back.domain.member.member.entity.Member;
-import com.back.domain.member.member.service.MemberService;
+import com.back.domain.member.entity.Member;
+import com.back.domain.member.service.MemberService;
 import com.back.domain.post.post.entity.Post;
 import com.back.domain.post.post.repository.PostRepository;
 import com.back.global.exception.ServiceException;
