+++ src/main/java/com/back/domain/member/common/MemberRole.java
@@ -1,4 +1,4 @@
-package com.back.domain.member.member.common;
+package com.back.domain.member.common;
 
 public enum MemberRole {
     USER,
