+++ src/main/java/com/back/domain/member/dto/MemberDto.java
@@ -1,6 +1,6 @@
-package com.back.domain.member.member.dto;
+package com.back.domain.member.dto;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 
 import java.time.LocalDateTime;
