+++ src/main/java/com/back/domain/member/dto/MemberLoginReqBody.java
@@ -1,4 +1,4 @@
-package com.back.domain.member.member.dto;
+package com.back.domain.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
