+++ src/main/java/com/back/domain/member/entity/Member.java
@@ -1,6 +1,6 @@
-package com.back.domain.member.member.entity;
+package com.back.domain.member.entity;
 
-import com.back.domain.member.member.common.MemberRole;
+import com.back.domain.member.common.MemberRole;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
