+++ src/main/java/com/back/domain/post/post/entity/PostFavorite.java
@@ -1,6 +1,6 @@
 package com.back.domain.post.post.entity;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
