+++ src/main/java/com/back/domain/report/report/entity/Report.java
@@ -1,6 +1,6 @@
 package com.back.domain.report.report.entity;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 import com.back.domain.report.report.common.ReportType;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.Entity;
