+++ src/main/java/com/back/domain/report/report/repository/ReportQueryRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
-import static com.back.domain.member.member.entity.QMember.member;
+import static com.back.domain.member.entity.QMember.member;
 import static com.back.domain.report.report.entity.QReport.report;
 
 @Repository
