+++ src/main/java/com/back/domain/reservation/entity/Reservation.java
@@ -1,6 +1,6 @@
 package com.back.domain.reservation.entity;
 
-import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.entity.Member;
 import com.back.domain.post.post.entity.Post;
 import com.back.domain.post.post.entity.PostOption;
 import com.back.domain.reservation.common.ReservationDeliveryMethod;
