+++ src/main/java/com/back/global/security/CustomUserDetailService.java
@@ -1,7 +1,7 @@
 package com.back.global.security;
 
-import com.back.domain.member.member.entity.Member;
-import com.back.domain.member.member.service.MemberService;
+import com.back.domain.member.entity.Member;
+import com.back.domain.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
