+++ src/test/java/com/back/domain/category/category/controller/CategoryAdmControllerTest.java
@@ -5,8 +5,8 @@ import com.back.domain.category.category.dto.CategoryCreateReqBody;
 import com.back.domain.category.category.dto.CategoryResBody;
 import com.back.domain.category.category.dto.CategoryUpdateReqBody;
 import com.back.domain.category.category.service.CategoryService;
-import com.back.domain.member.member.service.AuthTokenService;
-import com.back.domain.member.member.service.RefreshTokenStore;
+import com.back.domain.member.service.AuthTokenService;
+import com.back.domain.member.service.RefreshTokenStore;
 import com.back.global.security.SecurityUser;
 import com.back.global.web.CookieHelper;
 import com.fasterxml.jackson.databind.ObjectMapper;
