+++ src/test/java/com/back/domain/region/region/controller/RegionAdmControllerTest.java
@@ -1,7 +1,7 @@
 package com.back.domain.region.region.controller;
 
-import com.back.domain.member.member.service.AuthTokenService;
-import com.back.domain.member.member.service.RefreshTokenStore;
+import com.back.domain.member.service.AuthTokenService;
+import com.back.domain.member.service.RefreshTokenStore;
 import com.back.domain.region.region.dto.RegionCreateReqBody;
 import com.back.domain.region.region.dto.RegionResBody;
 import com.back.domain.region.region.dto.RegionUpdateReqBody;
